requirements.txt
chainlit==2.6.9
fastapi==0.116.1
httpx==0.28.1
humanize==4.12.3
langchain==0.3.27
langchain_community==0.3.27
langchain_core==0.3.74
langchain_huggingface==0.3.1
langchain_text_splitters==0.3.9
langgraph==0.6.5
pandas==2.3.1
protobuf>=5.26.0,<6.0
pydantic==2.11.7
python-dotenv==1.1.1
python_docx==1.2.0
uvicorn==0.35.0
google-generativeai
boto3==1.40.24